Negative dependencies are calculated using SCEV

+++ assignment_4/tests/test_dep.cpp
@@ -1,13 +1,13 @@
 int fun(int n, int arr[]) {
   int c = 10;
-  for (int i = 0; i < n; i++) {
+  for (long i = 0; i < n; i++) {
     c = i;
   }
-  for (int i = 0; i < n; i++) {
+  for (long i = 0; i < n; i++) {
     // int d = i + c;
     arr[i] = c * 2;
   }
-  for (int i = 0; i < n; i++) {
+  for (long i = 0; i < n; i++) {
     int d = arr[i + 2] + 3;
   }
   return 0;
